+++ src/cli/command-registrar-builder.js
@@ -9,6 +9,7 @@ import Create from './commands/public-cmds/create-cmd';
 import Export from './commands/public-cmds/export-cmd';
 import List from './commands/public-cmds/list-cmd';
 import Modify from './commands/public-cmds/modify-cmd';
+import Reset from './commands/public-cmds/reset-cmd';
 import Commit from './commands/public-cmds/commit-cmd';
 import Import from './commands/public-cmds/import-cmd';
 import ClearCache from './commands/public-cmds/clear-cache-cmd';
@@ -40,6 +41,7 @@ export default function registerCommands(): CommandRegistrar {
     new Export(),
     new Status(),
     new Modify(),
+    new Reset(),
     new List(),
     new Config(),
     new ClearCache(),